Fixed @ TECO binary
The @ TECO file from the Greenblatt "Stanford Cruft" tape was damaged.
This fixed file was contructed by disassembling @ TECO and comparing
against TECO 187 source code and another @ TECO from Samson's PS6
DECtape.

A file in octal notation is supplied.  Including lines starting with a
';', the first column exactly matches the damaged file.  Replacing the
bad words with the ones from the third column produces the fixed file;
lines with dashes are inserted, and ';' are removed.